# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ContentKit",
    "version": "v0",
    "description": "ContentKit's HTTP API, from the prefix the host mounts contentkit.Runtime.Handler at. Each module sits at its own sub-path (x-contentkit-modules); a host that mounts a module alone serves its routes beneath that mount instead. Generated from the route catalog (internal/httpapi) by `go generate ./internal/contract`; do not edit."
  },
  "paths": {
...
    "/{kind}/{id}/reaction": {
      "delete": {
        "tags": [
          "reactions"
        ],
        "summary": "Clears the caller's reaction to a target.",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionCounts"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {},
          {
            "host": []
          }
        ],
        "x-contentkit-module": "content",
        "x-contentkit-auth": "public",
        "x-contentkit-errors": [
          "forbidden",
          "invalid_request",
          "not_found",
          "rate_limited"
        ],
        "x-contentkit-error-sets": [
          "module:content",
          "request"
        ]
      },
      "get": {
        "tags": [
          "reactions"
        ],
        "summary": "A target's like and dislike counts and the caller's own reaction.",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionCounts"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {},
          {
            "host": []
          }
        ],
        "x-contentkit-module": "content",
        "x-contentkit-auth": "public",
        "x-contentkit-errors": [
          "not_found"
        ],
        "x-contentkit-error-sets": [
          "module:content",
          "request"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BanNotice": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "until": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "scope"
        ]
      },
...
      "ErrorDetails": {
        "type": "object",
        "properties": {
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "min_width": {
            "type": "integer"
          },
          "max_pixels": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "allowed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "size": {
            "type": "integer"
          },
          "max_bytes": {
            "type": "integer"
          },
          "frames": {
            "type": "integer"
          },
          "max_frames": {
            "type": "integer"
          },
          "seconds": {
            "type": "number"
          },
          "max_seconds": {
            "type": "number"
          }
        }
      },
      "ErrorReply": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "retry_after": {
            "type": "integer"
          },
          "action": {
            "type": "string",
            "enum": [
              "comment",
              "comment_reaction",
              "favorite",
              "poll_vote",
              "post_reaction",
              "reaction"
            ]
          },
          "ban": {
            "$ref": "#/components/schemas/BanNotice"
          },
          "blobs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "error",
          "code"
        ]
      },
...
      "ReactionCounts": {
        "type": "object",
        "properties": {
          "likes": {
            "type": "integer"
          },
          "dislikes": {
            "type": "integer"
          },
          "mine": {
            "type": "integer"
          }
        },
        "required": [
          "likes",
          "dislikes",
          "mine"
        ]
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "animation_not_allowed",
          "animation_too_long",
          "animation_unsupported",
          "checksum_mismatch",
          "comment_banned",
          "conflict",
          "forbidden",
          "gone",
          "image_too_large",
          "image_too_small",
          "image_unreadable",
          "incomplete",
          "internal_error",
          "invalid_request",
          "moderation_rejected",
          "not_configured",
          "not_found",
          "not_uploaded",
          "quota_exceeded",
          "rate_limited",
          "tenant_mismatch",
          "too_large",
          "too_many_files",
          "type_not_allowed",
          "unauthorized",
          "unavailable",
          "video_over_budget",
          "video_too_large",
          "video_too_long"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "A refusal: a code of the operation's x-contentkit-errors or of an x-contentkit-error-sets set it names.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorReply"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "host": {
        "type": "http",
        "scheme": "bearer",
        "description": "Whatever the host's auth middleware accepts: ContentKit reads the actor it puts in the request context."
      }
    }
  },
...
}